refactor(enum): SeatStatus를 좌석 조회 모델로 이동
