+++ core/core-api/src/main/java/com/ticket/core/domain/performanceseat/query/model/SeatStatus.java
@@ -1,8 +1,8 @@
-package com.ticket.core.enums;
+package com.ticket.core.domain.performanceseat.query.model;
 
 /**
  * 좌석 상태 API 응답 전용 enum.
- * DB 저장용 PerformanceSeatState와 분리하여, 클라이언트에게는 단순한 2가지 상태만 노출합니다.
+ * DB 저장용 PerformanceSeatState와 분리하여, 클라이언트에는 단순한 2가지 상태만 노출한다.
  */
 public enum SeatStatus {
     AVAILABLE("선택 가능"),
